Put some transform stuff in its own .cpp file. Fixed some warnings for gcc-4.3.
git-svn-id: https://spacegameengine.svn.sourceforge.net/svnroot/spacegameengine@512 469afb1b-a615-0410-ae09-9c7fe5ec7362

+++ src/plugins/opengl/src/shader.cpp
@@ -25,7 +25,7 @@ sge::ogl::shader::shader(const GLenum type, const std::string& source)
 : id_(glCreateShader(type))
 {
 	const char* const ptr = source.c_str();
-	const int len = source.size();
+	const int len = static_cast<int>(source.size());
 	glShaderSource(id(), 1, const_cast<const char**>(&ptr), &len);
 
 	glCompileShader(id());
